+++ lib/src/widgets/card_cvc_form_field.dart
@@ -19,7 +19,8 @@ class CardCvcFormField extends StatelessWidget {
   final String Function(String) validator;
 
   static const defaultErrorText = "Invalid CVV";
-  static const defaultDecoration = InputDecoration(border: OutlineInputBorder(), labelText: "CVV", hintText: "XXX");
+  static const defaultDecoration = InputDecoration(
+      border: OutlineInputBorder(), labelText: "CVV", hintText: "XXX");
 
   @override
   Widget build(BuildContext context) {