+++ lib/src/widgets/card_form.dart
@@ -54,9 +54,11 @@ class _CardFormState extends State<CardForm> {
               onChanged: (number) => _validationModel.number = number,
               validator: (text) => _validationModel.validateNumber()
                   ? null
-                  : widget.cardNumberErrorText ?? CardNumberFormField.defaultErrorText,
+                  : widget.cardNumberErrorText ??
+                      CardNumberFormField.defaultErrorText,
               onSaved: (text) => widget.card.number = text,
-              decoration: widget.cardNumberDecoration ?? CardNumberFormField.defaultDecoration,
+              decoration: widget.cardNumberDecoration ??
+                  CardNumberFormField.defaultDecoration,
             ),
           ),
           Container(
@@ -75,8 +77,10 @@ class _CardFormState extends State<CardForm> {
                 },
                 validator: (text) => _validationModel.validateDate()
                     ? null
-                    : widget.cardExpiryErrorText ?? CardExpiryFormField.defaultErrorText,
-                decoration: widget.cardExpiryDecoration ?? CardExpiryFormField.defaultDecoration,
+                    : widget.cardExpiryErrorText ??
+                        CardExpiryFormField.defaultErrorText,
+                decoration: widget.cardExpiryDecoration ??
+                    CardExpiryFormField.defaultDecoration,
               )),
           Container(
             padding: const EdgeInsets.symmetric(vertical: 8.0),
@@ -85,9 +89,12 @@ class _CardFormState extends State<CardForm> {
               initialValue: _validationModel.cvc ?? widget.card.cvc,
               onChanged: (text) => _validationModel.cvc = text,
               onSaved: (text) => widget.card.cvc = text,
-              validator: (text) =>
-                  _validationModel.validateCVC() ? null : widget.cardCvcErrorText ?? CardCvcFormField.defaultErrorText,
-              decoration: widget.cardCvcDecoration ?? CardCvcFormField.defaultDecoration,
+              validator: (text) => _validationModel.validateCVC()
+                  ? null
+                  : widget.cardCvcErrorText ??
+                      CardCvcFormField.defaultErrorText,
+              decoration: widget.cardCvcDecoration ??
+                  CardCvcFormField.defaultDecoration,
             ),
           ),
         ],