+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Stripe SDK Changelog
 
+## 2.4.2
+
+* Allow custom form field error text
+* Fix bug where card number input field would allow more than 16 digits
+
 ## 2.4.1
 
 * Add support for custom input decorators on card form and form fields